+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/inportalsurvey/FeedBackResponse.java
@@ -1,4 +1,5 @@
-package org.pucar.dristi.caselifecycle.inportalsurvey.internal.web.models;
+// HAND-CURATED — lifted by Phase 35 (contract-lift)
+package org.pucar.dristi.common.contract.inportalsurvey;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
@@ -10,7 +11,7 @@ import org.egov.common.contract.response.ResponseInfo;
 import org.springframework.validation.annotation.Validated;
 
 /**
- * EligibilityResponse
+ * FeedBackResponse
  */
 @Validated
 @jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2025-10-14T19:19:54.104875784+05:30[Asia/Kolkata]")
@@ -18,14 +19,14 @@ import org.springframework.validation.annotation.Validated;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class EligibilityResponse {
+public class FeedBackResponse {
 
-    @JsonProperty("ResponseInfo")
+    @JsonProperty("responseInfo")
     @Valid
     private ResponseInfo responseInfo;
 
-    @JsonProperty("Eligibility")
+    @JsonProperty("feedback")
     @Valid
-    private Eligibility eligibility;
+    private FeedBack feedback;
 
 }
