+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/inportalsurvey/RemindMeLaterRequest.java
@@ -1,17 +1,17 @@
-package org.pucar.dristi.caselifecycle.inportalsurvey.internal.web.models;
+// HAND-CURATED — lifted by Phase 35 (contract-lift)
+package org.pucar.dristi.common.contract.inportalsurvey;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotNull;
-import org.egov.common.contract.request.RequestInfo;
-import org.springframework.validation.annotation.Validated;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import lombok.Data;
 import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.egov.common.contract.request.RequestInfo;
+import org.springframework.validation.annotation.Validated;
 
 /**
- * EligibilityRequest
+ * RemindMeLater
  */
 @Validated
 @jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2025-10-14T19:19:54.104875784+05:30[Asia/Kolkata]")
@@ -19,11 +19,10 @@ import lombok.Builder;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class EligibilityRequest {
+public class RemindMeLaterRequest {
 
     @JsonProperty("RequestInfo")
     @Valid
-    @NotNull
-    private RequestInfo requestInfo = null;
+    private RequestInfo requestInfo;
 
 }
