+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/inportalsurvey/SurveyTrackerRequest.java
@@ -1,15 +1,18 @@
-package org.pucar.dristi.caselifecycle.inportalsurvey.internal.web.models;
+// HAND-CURATED — lifted by Phase 35 (contract-lift)
+package org.pucar.dristi.common.contract.inportalsurvey;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.egov.common.contract.request.RequestInfo;
 import org.springframework.validation.annotation.Validated;
 
 /**
- * Eligibility
+ * SurveyTrackerRequest
  */
 @Validated
 @jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2025-10-14T19:19:54.104875784+05:30[Asia/Kolkata]")
@@ -17,10 +20,16 @@ import org.springframework.validation.annotation.Validated;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class Eligibility {
+public class SurveyTrackerRequest {
 
-    @JsonProperty("isEligible")
+    @JsonProperty("RequestInfo")
     @Valid
-    private Boolean isEligible;
+    @NotNull
+    private RequestInfo requestInfo = null;
+
+    @JsonProperty("surveyTracker")
+    @Valid
+    @NotNull
+    private SurveyTracker surveyTracker = null;
 
 }
