+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/inportalsurvey/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Inportalsurvey subdomain's contract DTOs — request/response envelopes plus their
+ * transitive payload types, lifted by Phase 35 of the per-module migration
+ * pipeline. Exposed as a {@link org.springframework.modulith.NamedInterface}
+ * so callers in other modules (caselifecycle, etc.) can depend on these
+ * types without crossing the dristi-common boundary.
+ */
+@org.springframework.modulith.NamedInterface("contract-inportalsurvey")
+package org.pucar.dristi.common.contract.inportalsurvey;
